Global profile for login, logout for GIB users; watchers no longer necessary

+++ new-wetty/bash_logout.sh
@@ -1,9 +1,12 @@
 # Executes when the user logs out
 
 
-if [ $USER != "Empty" ]; then
-    # Free instance
-    curl http://$MANAGER_NODE:5000/api/instance/freeme
-    # Deletes all user files that can be deleted
-    rm /home/ipt/*
+if [ $USER != "root" ]; then
+
+    if [ $USER != "Empty" ]; then
+        # Free instance
+        curl http://$MANAGER_NODE:5000/api/instance/freeme
+        # Deletes all user files that can be deleted
+        rm /home/ipt/*
+    fi
 fi
